pyproject.toml
[project]
name = "universal-credit-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.121.3",
    "fitz>=0.0.1.dev2",
    "google-genai>=1.52.0",
    "langchain>=1.0.8",
    "langchain-google-genai>=3.1.0",
    "openai>=2.8.1",
    "pydantic>=2.12.4",
    "pymupdf>=1.26.6",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "streamlit>=1.51.0",
    "uvicorn>=0.38.0",
]
